+++ solc/CommandLineInterface.cpp
@@ -36,6 +36,8 @@
 #include <libsolidity/interface/DebugSettings.h>
 #include <libsolidity/interface/ImportRemapper.h>
 #include <libsolidity/interface/StorageLayout.h>
+#include <libsolidity/lsp/LanguageServer.h>
+#include <libsolidity/lsp/Transport.h>
 
 #include <libyul/AssemblyStack.h>
 
@@ -54,6 +56,7 @@
 #include <libsolutil/JSON.h>
 
 #include <algorithm>
+#include <fstream>
 #include <memory>
 
 #include <range/v3/view/map.hpp>
@@ -507,7 +510,11 @@ bool CommandLineInterface::readInputFiles()
 			m_fileReader.setStdin(readUntilEnd(m_sin));
 	}
 
-	if (m_fileReader.sourceCodes().empty() && !m_standardJsonInput.has_value())
+	if (
+		m_options.input.mode != InputMode::LanguageServer &&
+		m_fileReader.sourceCodes().empty() &&
+		!m_standardJsonInput.has_value()
+	)
 	{
 		serr() << "All specified input files either do not exist or are not regular files." << endl;
 		return false;
@@ -623,6 +630,9 @@ bool CommandLineInterface::processInput()
 		m_standardJsonInput.reset();
 		break;
 	}
+	case InputMode::LanguageServer:
+		serveLSP();
+		break;
 	case InputMode::Assembler:
 		if (!assemble(m_options.assembly.inputLanguage, m_options.assembly.targetMachine))
 			return false;
@@ -748,7 +758,7 @@ bool CommandLineInterface::compile()
 	catch (Error const& _error)
 	{
 		if (_error.type() == Error::Type::DocstringParsingError)
-			serr() << "Documentation parsing error: " << *boost::get_error_info<errinfo_comment>(_error) << endl;
+			serr() << "Documentation parsing error: " << *_error.comment() << endl;
 		else
 		{
 			m_hasOutput = true;
@@ -885,6 +895,12 @@ void CommandLineInterface::handleAst()
 			ASTJsonConverter(m_compiler->state(), m_compiler->sourceIndices()).print(sout(), m_compiler->ast(sourceCode.first));
 		}
 	}
+}
+
+void CommandLineInterface::serveLSP()
+{
+	lsp::JSONTransport transport;
+	lsp::LanguageServer{transport}.run();
 }
 
 bool CommandLineInterface::link()