+++ Changelog.md
@@ -4,6 +4,7 @@ Language Features:
 
 
 Compiler Features:
+ * Commandline Interface: Add ``--lsp`` option to get ``solc`` act as a Language Server (LSP) communicationg over stdio.
 
 
 Bugfixes: